% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_box}
\alias{calculate_box}
\title{Calculate docking box parameters from PDB structures}
\usage{
calculate_box(pdb_files, padding = 5)
}
\arguments{
\item{pdb_files}{Character vector of paths to PDB files}

\item{padding}{Numeric, additional padding to add around the protein in Angstroms, default 5}
}
\value{
Data frame containing file names, center coordinates (x,y,z) and box dimensions (x,y,z)
}
\description{
This function analyzes PDB files to determine appropriate docking box
dimensions and center coordinates based on the protein structure.
}
\examples{
\dontrun{
box_params <- calculate_box(pdb_files, padding = 4)
}

}
