% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_alt.R
\name{collect_pdb_files}
\alias{collect_pdb_files}
\title{Collect PDB files from input paths}
\usage{
collect_pdb_files(inputs)
}
\arguments{
\item{inputs}{Character vector of file paths or directories}
}
\value{
Character vector of PDB file paths
}
\description{
Collect PDB files from input paths
}
