% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_ligand}
\alias{download_ligand}
\title{Download ligand structures from PubChem}
\usage{
download_ligand(cids, out_dir = "ligands", verify_ssl = TRUE)
}
\arguments{
\item{cids}{Character vector of PubChem Compound IDs (CIDs) to download}

\item{out_dir}{Character, output directory for downloaded files, default "ligands"}

\item{verify_ssl}{Logical, whether to verify SSL certificates, default TRUE}
}
\value{
List of successfully downloaded file paths
}
\description{
This function downloads chemical compound structures from PubChem database using their CIDs.
}
\examples{
\dontrun{
files <- download_ligand(c("2244", "5090"), out_dir = "compounds")
}

}
