% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_receptor}
\alias{download_receptor}
\title{Download PDB files from RCSB PDB database}
\usage{
download_receptor(pdb_ids, out_dir = "./", verify_ssl = TRUE)
}
\arguments{
\item{pdb_ids}{Character vector of PDB IDs to download}

\item{out_dir}{Character, output directory for downloaded files, default "./"}

\item{verify_ssl}{Logical, whether to verify SSL certificates, default TRUE}
}
\value{
List of successfully downloaded file paths
}
\description{
This function downloads PDB files from the RCSB Protein Data Bank using their PDB IDs.
}
\examples{
\dontrun{
files <- download_receptor(c("1iep", "4hg7"), out_dir = "receptors")
}

}
