% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_alt.R
\name{has_alternate_locations}
\alias{has_alternate_locations}
\title{Check if PDB file contains alternate locations}
\usage{
has_alternate_locations(pdb_file)
}
\arguments{
\item{pdb_file}{Path to PDB file}
}
\value{
Logical indicating if alternate locations exist
}
\description{
Check if PDB file contains alternate locations
}
\keyword{internal}
