% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_alt.R
\name{split_pdb_with_script}
\alias{split_pdb_with_script}
\title{Split PDB file with external Python script}
\usage{
split_pdb_with_script(
  pdb_file,
  out_dir,
  stem,
  python_path,
  script_path,
  out_file,
  keep_label
)
}
\arguments{
\item{pdb_file}{Path to PDB file}

\item{out_dir}{Output directory}

\item{stem}{File stem name}

\item{python_path}{Path to Python executable}

\item{script_path}{Path to splitting script}

\item{out_file}{Expected output file path}

\item{keep_label}{Label to keep}
}
\value{
Character vector with path to processed file if successful
}
\description{
Split PDB file with external Python script
}
