% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_receptor.R
\name{trim_receptor}
\alias{trim_receptor}
\title{Trim receptor PDB files to keep only protein atoms}
\usage{
trim_receptor(input_paths, output_dir = "trimmed")
}
\arguments{
\item{input_paths}{Character vector of file paths or directories containing PDB files}

\item{output_dir}{Character string specifying the output directory, default is "trimmed"}
}
\value{
Character vector of paths to the trimmed PDB files
}
\description{
Trim receptor PDB files to keep only protein atoms
}
\examples{
\dontrun{
trim_receptor("./test/1a7c.pdb")
trim_receptor(c("./test/1a7c.pdb", "./test/1abc.pdb"), output_dir = "my_output")
trim_receptor("./test", output_dir = "processed")
}
}
