% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receptor.R
\name{write_configs}
\alias{write_configs}
\title{Generate qvina or qvina-w config files for docking}
\usage{
write_configs(
  receptor_paths,
  ligand_paths,
  box_df,
  out_dir = "configs",
  exhaustiveness = 8
)
}
\arguments{
\item{receptor_paths}{Character vector, full paths to receptor PDBQT files}

\item{ligand_paths}{Character vector, full paths to ligand PDBQT files}

\item{box_df}{Data.frame generated by calculate_box(), must contain columns:
file (basename without path), center_x/center_y/center_z,
size_x/size_y/size_z}

\item{out_dir}{Output directory for configuration files}

\item{exhaustiveness}{Integer, search intensity, default 8}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{          All written configuration file paths (invisible return)
}\if{html}{\out{</div>}}
}
\description{
Generate qvina or qvina-w config files for docking
}
\examples{
\dontrun{
# Create configuration files for docking
config_files <- write_configs(
  receptor_paths = receptor_pdbqt,
  ligand_paths = ligand_pdbqt,
  box_df = box_params,
  out_dir = "configs"
)
}
}
